+++ examples/main/modules.tf
@@ -1,0 +1,89 @@
+module "azure_region" {
+  source  = "claranet/regions/azurerm"
+  version = "x.x.x"
+
+  azure_region = var.azure_region
+}
+
+module "rg" {
+  source  = "claranet/rg/azurerm"
+  version = "x.x.x"
+
+  location    = module.azure_region.location
+  client_name = var.client_name
+  environment = var.environment
+  stack       = var.stack
+}
+
+module "logs" {
+  source  = "claranet/run-common/azurerm//modules/logs"
+  version = "x.x.x"
+
+  client_name         = var.client_name
+  environment         = var.environment
+  stack               = var.stack
+  location            = module.azure_region.location
+  location_short      = module.azure_region.location_short
+  resource_group_name = module.rg.resource_group_name
+}
+
+module "storage_account" {
+  source  = "claranet/storage-account/azurerm"
+  version = "x.x.x"
+
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
+  client_name    = var.client_name
+  environment    = var.environment
+  stack          = var.stack
+
+  resource_group_name = module.rg.resource_group_name
+
+  account_replication_type = "LRS"
+
+  storage_blob_data_protection = {
+    change_feed_enabled                       = true
+    versioning_enabled                        = true
+    delete_retention_policy_in_days           = 42
+    container_delete_retention_policy_in_days = 42
+    container_point_in_time_restore           = true
+  }
+
+  logs_destinations_ids = [
+    module.logs.logs_storage_account_id,
+    module.logs.log_analytics_workspace_id
+  ]
+
+  containers = [
+    {
+      name = "bloc1"
+    },
+    {
+      name                  = "bloc2"
+      container_access_type = "blob"
+    }
+  ]
+
+  file_shares = [
+    {
+      name        = "share1smb"
+      quota_in_gb = 50
+    }
+  ]
+
+  tables = [
+    {
+      name = "table1"
+    }
+  ]
+
+  queues = [
+    {
+      name = "mystoragequeue"
+    }
+  ]
+
+  extra_tags = {
+    foo = "bar"
+  }
+}