+++ variables.tf
@@ -1,110 +1,232 @@
-variable "client_name" {
-  description = "Name of client"
-  type        = string
-}
+# Common
 
 variable "location_short" {
-  description = "Short string for Azure location."
+  description = "Short string for Azure location"
   type        = string
 }
 
 variable "location" {
+  description = "Azure location"
+  type        = string
 }
 
-variable "resource_group_name" {
+variable "client_name" {
+  description = "Client name/account used in naming"
+  type        = string
 }
 
 variable "environment" {
+  description = "Project environment"
+  type        = string
 }
 
 variable "stack" {
+  description = "Project stack name"
+  type        = string
 }
 
-variable "storage_account_name" {
+variable "resource_group_name" {
+  description = "Resource group name"
+  type        = string
+}
+
+# Storage account parameters
+
+variable "account_kind" {
+  description = "Defines the Kind of account. Valid options are `BlobStorage`, `BlockBlobStorage`, `FileStorage`, `Storage` and `StorageV2`. Changing this forces a new resource to be created. Defaults to StorageV2."
+  type        = string
+  default     = "StorageV2"
 }
 
 variable "account_tier" {
+  description = "Defines the Tier to use for this storage account. Valid options are `Standard` and `Premium`. For `BlockBlobStorage` and `FileStorage` accounts only `Premium` is valid. Changing this forces a new resource to be created."
+  type        = string
+  default     = "Standard"
 }
 
 variable "access_tier" {
-  type    = string
-  default = "Hot"
+  description = "Defines the access tier for `BlobStorage`, `FileStorage` and `StorageV2` accounts. Valid options are `Hot` and `Cool`, defaults to `Hot`."
+  type        = string
+  default     = "Hot"
 }
 
-variable "account_kind" {
-  type    = string
-  default = "StorageV2"
+variable "account_replication_type" {
+  description = "Defines the type of replication to use for this storage account. Valid options are `LRS`, `GRS`, `RAGRS`, `ZRS`, `GZRS` and `RAGZRS`."
+  type        = string
+  default     = "ZRS"
 }
 
-variable "replication_type" {
+variable "https_traffic_only_enabled" {
+  description = "Boolean flag which forces HTTPS if enabled."
+  type        = bool
+  default     = true
 }
 
-variable "extra_tags" {
-  type    = map(string)
-  default = {}
+variable "min_tls_version" {
+  description = "The minimum supported TLS version for the storage account. Possible values are `TLS1_0`, `TLS1_1`, and `TLS1_2`. "
+  type        = string
+  default     = "TLS1_2"
 }
 
-variable "enable_https_traffic_only" {
-  type    = string
-  default = "true"
+variable "public_nested_items_allowed" {
+  description = "Allow or disallow nested items within this Account to opt into being public."
+  type        = bool
+  default     = false
 }
 
 variable "custom_domain_name" {
-  type    = string
-  default = ""
-}
-
-variable "name_prefix" {
-  description = "Optional prefix for storage account name"
+  description = "The Custom Domain Name to use for the Storage Account, which will be validated by Azure."
   type        = string
-  default     = ""
-}
-
-variable "container_access_type" {
-  description = "he Access Level configured for this Container. Possible values are blob, container or private"
-  type        = string
-  default     = "private"
-}
-
-variable "static_website_config" {
-  type    = list(map(string))
-  default = []
+  default     = null
 }
 
 variable "use_subdomain" {
-  type    = bool
-  default = false
+  description = "Should the Custom Domain Name be validated by using indirect CNAME validation?"
+  type        = bool
+  default     = false
+}
+
+variable "static_website_config" {
+  description = "Static website configuration. Can only be set when the `account_kind` is set to `StorageV2` or `BlockBlobStorage`."
+  type = object({
+    index_document     = optional(string)
+    error_404_document = optional(string)
+  })
+  default = null
+}
+
+variable "shared_access_key_enabled" {
+  description = "Indicates whether the storage account permits requests to be authorized with the account access key via Shared Key. If false, then all requests, including shared access signatures, must be authorized with Azure Active Directory (Azure AD)."
+  type        = bool
+  default     = true
+}
+
+variable "nfsv3_enabled" {
+  description = "Is NFSv3 protocol enabled? Changing this forces a new resource to be created."
+  type        = bool
+  default     = false
+}
+
+# Identity
+variable "identity_type" {
+  description = "Specifies the type of Managed Service Identity that should be configured on this Storage Account. Possible values are `SystemAssigned`, `UserAssigned`, `SystemAssigned, UserAssigned` (to enable both)."
+  type        = string
+  default     = "SystemAssigned"
+}
+
+variable "identity_ids" {
+  description = "Specifies a list of User Assigned Managed Identity IDs to be assigned to this Storage Account."
+  type        = list(string)
+  default     = null
+}
+
+# Storage Firewall
+
+variable "network_rules_enabled" {
+  description = "Boolean to enable Network Rules on the Storage Account, requires `network_bypass`, `ip_rules`, `subnet_ids` or `default_firewall_action` correctly set if enabled."
+  type        = bool
+  default     = true
 }
 
 variable "network_bypass" {
-  description = "Specifies whether traffic is bypassed for 'Logging', 'Metrics', 'AzureServices' or 'None'"
+  description = "Specifies whether traffic is bypassed for 'Logging', 'Metrics', 'AzureServices' or 'None'."
   type        = list(string)
   default     = ["None"]
 }
 
-variable "ip_rules" {
-  description = "IPs to allow access to that storage account"
-  type        = list(string)
-  default     = []
-}
-
-variable "containers" {
-  description = "Containers to create on the storage account"
+variable "allowed_cidrs" {
+  description = "List of CIDR to allow access to that storage account."
   type        = list(string)
   default     = []
 }
 
 variable "subnet_ids" {
-  description = "Subnets to allow access to that storage account"
+  description = "Subnets to allow access to that storage account."
   type        = list(string)
   default     = []
 }
 
 variable "default_firewall_action" {
-  description = "Which default firewalling policy to apply. Valid values are Allow or Deny"
+  description = "Which default firewalling policy to apply. Valid values are `Allow` or `Deny`."
+  type        = string
   default     = "Deny"
 }
 
-variable "min_tls_version" {
-  default = "TLS1_2"
+# Data protection
+
+variable "storage_blob_data_protection" {
+  description = "Storage account blob Data protection parameters."
+  type = object({
+    change_feed_enabled                       = bool
+    versioning_enabled                        = bool
+    delete_retention_policy_in_days           = number
+    container_delete_retention_policy_in_days = number
+    container_point_in_time_restore           = bool
+  })
+  default = {
+    change_feed_enabled                       = true
+    versioning_enabled                        = true
+    delete_retention_policy_in_days           = 30
+    container_delete_retention_policy_in_days = 30
+    container_point_in_time_restore           = true
+  }
+}
+
+# Threat protection
+variable "advanced_threat_protection_enabled" {
+  description = "Boolean flag which controls if advanced threat protection is enabled, see [documentation](https://docs.microsoft.com/en-us/azure/storage/common/storage-advanced-threat-protection?tabs=azure-portal) for more information."
+  type        = bool
+  default     = false
+}
+
+# Data creation/bootstrap
+
+variable "containers" {
+  description = "List of objects to create some Blob containers in this Storage Account."
+  type = list(object({
+    name                  = string
+    container_access_type = optional(string)
+    metadata              = optional(map(string))
+  }))
+  default = []
+}
+
+variable "file_shares" {
+  description = "List of objects to create some File Shares in this Storage Account."
+  type = list(object({
+    name             = string
+    quota_in_gb      = number
+    enabled_protocol = optional(string)
+    metadata         = optional(map(string))
+    acl = optional(list(object({
+      id          = string
+      permissions = string
+      start       = optional(string)
+      expiry      = optional(string)
+    })))
+  }))
+  default = []
+}
+
+variable "tables" {
+  description = "List of objects to create some Tables in this Storage Account."
+  type = list(object({
+    name = string
+    acl = optional(list(object({
+      id          = string
+      permissions = string
+      start       = optional(string)
+      expiry      = optional(string)
+    })))
+  }))
+  default = []
+}
+
+variable "queues" {
+  description = "List of objects to create some Queues in this Storage Account."
+  type = list(object({
+    name     = string
+    metadata = optional(map(string))
+  }))
+  default = []
 }