+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+# Unreleased
+
 Added
-  * AZ-130: First Release
+  * AZ-130: Module Azure storage account - first release
+  * AZ-504: Adding diagnostic logging
+  * AZ-515: Compatible with CAF Naming
+  * AZ-615: Add an option to enable or disable default tags
+  * AZ-717: Compatible with Terraform 1+ and AzureRM provider v3+